# backend/open-shop-Api-spec/open-shop-api-spec.yaml
openapi: 3.0.3
info:
  title: Open Shop API
  description: E-commerce platform API for Open Shop application
  version: 1.0.0
  contact:
    name: Open Shop API Support
    email: api-support@openshop.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.openshop.com/v1
    description: Production server
  - url: https://staging-api.openshop.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
security:
  - bearerAuth: [ ]
paths:
...
  /products:
    get:
      tags:
        - product
      summary: List products with filtering and pagination
      description: |
        Retrieve a list of products with optional filtering and sorting.
        Supports filtering by category, price range, availability, and other attributes.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last product in the returned list.

        Rate limited to 100 requests per minute per IP address.
      operationId: listProducts
      security: [ ]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of products to return per request (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: category
          in: query
          description: Filter by category ID
          schema:
            $ref: '#/components/schemas/ID'
        - name: minPrice
          in: query
          description: Filter by minimum price
          schema:
            type: number
            format: float
            minimum: 0
            example: 10.00
        - name: maxPrice
          in: query
          description: Filter by maximum price
          schema:
            type: number
            format: float
            minimum: 0
            example: 100.00
        - name: inStock
          in: query
          description: Filter by availability
          schema:
            type: boolean
            example: true
        - name: sortBy
          in: query
          description: Sort products by field
          schema:
            type: string
            enum: [ name, price, createdAt, rating, popularity ]
            default: createdAt
            example: price
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
            example: asc
        - name: search
          in: query
          description: Search term for product name or description
          schema:
            type: string
            minLength: 1
            maxLength: 100
            example: wireless headphones
      responses:
        '200':
          description: Products retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                type: object
                required: [ products, timestamp ]
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  timestamp:
                    $ref: '#/components/schemas/Timestamp'
              examples:
                product_list:
                  $ref: '#/components/examples/product_list'
  /products/{id}:
    get:
      tags:
        - product
      summary: Get single product details
      description: |
        Retrieve detailed information about a specific product including images, specifications,
        availability, and related products.

        Rate limited to 200 requests per minute per IP address.
      operationId: getProduct
      security: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Product identifier
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Product details retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                product_details:
                  $ref: '#/components/examples/product_detail'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/search:
    get:
      tags:
        - product
      summary: Search products
      description: |
        Full-text search for products by name, description, tags, and specifications.
        Supports advanced filtering and sorting of search results.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last product in the returned list.

        Rate limited to 50 requests per minute per IP address.
      operationId: searchProducts
      security: [ ]
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
            maxLength: 200
            example: wireless headphones Bluetooth
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of results to return per request (default 20, max 50)
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
            example: 20
        - name: category
          in: query
          description: Filter by category ID
          schema:
            $ref: '#/components/schemas/ID'
        - name: minPrice
          in: query
          description: Filter by minimum price
          schema:
            type: number
            format: float
            minimum: 0
            example: 10.00
        - name: maxPrice
          in: query
          description: Filter by maximum price
          schema:
            type: number
            format: float
            minimum: 0
            example: 100.00
        - name: sortBy
          in: query
          description: Sort results by field
          schema:
            type: string
            enum: [ relevance, name, price, rating, createdAt ]
            default: relevance
            example: relevance
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
            example: desc
      responses:
        '200':
          description: Search results retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSearchResponse'
  /categories:
    get:
      tags:
        - product
      summary: List all categories
      description: |
        Retrieve all product categories in a hierarchical structure.
        Includes category metadata and product counts.

        Rate limited to 100 requests per minute per IP address.
      operationId: listCategories
      security: [ ]
      parameters:
        - name: includeEmpty
          in: query
          description: Include categories with no products
          schema:
            type: boolean
            default: false
            example: true
      responses:
        '200':
          description: Categories retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              examples:
                categories_list:
                  $ref: '#/components/examples/categories_list'
  /categories/{id}:
    get:
      tags:
        - product
      summary: Get category details
      description: |
        Retrieve detailed information about a specific category including
        subcategories and metadata.

        Rate limited to 100 requests per minute per IP address.
      operationId: getCategory
      security: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Category identifier
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Category details retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/NotFound'
  /categories/{id}/products:
    get:
      tags:
        - product
      summary: Get products in a category
      description: |
        Retrieve all products belonging to a specific category with filtering.
        Includes products from subcategories if specified.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last product in the returned list.

        Rate limited to 100 requests per minute per IP address.
      operationId: getCategoryProducts
      security: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Category identifier
          schema:
            $ref: '#/components/schemas/ID'
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of products to return per request (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: includeSubcategories
          in: query
          description: Include products from subcategories
          schema:
            type: boolean
            default: false
            example: true
        - name: minPrice
          in: query
          description: Filter by minimum price
          schema:
            type: number
            format: float
            minimum: 0
            example: 10.00
        - name: maxPrice
          in: query
          description: Filter by maximum price
          schema:
            type: number
            format: float
            minimum: 0
            example: 100.00
        - name: inStock
          in: query
          description: Filter by availability
          schema:
            type: boolean
            example: true
        - name: sortBy
          in: query
          description: Sort products by field
          schema:
            type: string
            enum: [ name, price, createdAt, rating, popularity ]
            default: createdAt
            example: price
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
            example: asc
      responses:
        '200':
          description: Category products retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryProductsResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # Shopping Cart Endpoints
  /cart:
    get:
      tags:
        - product
      summary: Get current cart
      description: |
        Retrieve the authenticated customer's current shopping cart with all items,
        pricing information, and applicable discounts.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 100 requests per minute per user.
      operationId: getCart
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Cart retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
              examples:
                cart_with_items:
                  $ref: '#/components/examples/cart_with_items'
                empty_cart:
                  $ref: '#/components/examples/empty_cart'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - product
      summary: Clear entire cart
      description: |
        Remove all items from the authenticated customer's shopping cart.
        This action cannot be undone.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 20 requests per minute per user.
      operationId: clearCart
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Cart cleared successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
              examples:
                cart_cleared:
                  $ref: '#/components/examples/cart_cleared'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cart/items:
    post:
      tags:
        - product
      summary: Add item to cart
      description: |
        Add a product to the authenticated customer's shopping cart.
        If the product is already in the cart, the quantity will be updated.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 50 requests per minute per user.
      operationId: addCartItem
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
            examples:
              add_product:
                $ref: '#/components/examples/cart_add_item'
      responses:
        '201':
          description: Item added to cart successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartActionResponse'
              examples:
                item_added:
                  $ref: '#/components/examples/item_added'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                product_not_found:
                  $ref: '#/components/examples/product_not_found'
        '422':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /cart/items/{itemId}:
    put:
      tags:
        - product
      summary: Update cart item quantity
      description: |
        Update the quantity of a specific item in the authenticated customer's cart.
        Setting quantity to 0 will remove the item from the cart.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 50 requests per minute per user.
      operationId: updateCartItem
      security:
        - bearerAuth: [ ]
      parameters:
        - name: itemId
          in: path
          required: true
          description: Cart item identifier
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
            examples:
              update_quantity:
                $ref: '#/components/examples/update_quantity'
      responses:
        '200':
          description: Cart item updated successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                item_not_found:
                  $ref: '#/components/examples/item_not_found'
    delete:
      tags:
        - product
      summary: Remove item from cart
      description: |
        Remove a specific item from the authenticated customer's shopping cart.
        This action cannot be undone.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 50 requests per minute per user.
      operationId: removeCartItem
      security:
        - bearerAuth: [ ]
      parameters:
        - name: itemId
          in: path
          required: true
          description: Cart item identifier
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Item removed from cart successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartActionResponse'
              examples:
                item_removed:
                  $ref: '#/components/examples/item_removed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Cart item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                item_not_found:
                  $ref: '#/components/examples/item_not_found'
  # Order Management Endpoints
  /orders:
    get:
      tags:
        - product
      summary: List customer orders
      description: |
        Retrieve orders for the authenticated customer.
        Supports filtering by status and date range with sorting options.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last order in the returned list.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 100 requests per minute per user.
      operationId: listOrders
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of orders to return per request (default 20, max 50)
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
            example: 20
        - name: status
          in: query
          description: Filter by order status
          schema:
            type: string
            enum: [ pending, confirmed, processing, shipped, delivered, cancelled ]
            example: confirmed
        - name: startDate
          in: query
          description: Filter orders from this date (inclusive)
          schema:
            type: string
            example: 2024-01-01
        - name: endDate
          in: query
          description: Filter orders until this date (inclusive)
          schema:
            type: string
            example: 2024-01-31
        - name: sortBy
          in: query
          description: Sort orders by field
          schema:
            type: string
            enum: [ createdAt, updatedAt, totalAmount, orderNumber ]
            default: createdAt
            example: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
            example: desc
      responses:
        '200':
          description: Orders retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                type: object
                required: [ orders, timestamp ]
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  timestamp:
                    $ref: '#/components/schemas/Timestamp'
              examples:
                orders_list:
                  $ref: '#/components/examples/orders_list'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - product
      summary: Create new order
      description: |
        Create a new order from the authenticated customer's cart items and initialize payment intent.

        This endpoint will:
        - Create an order with status 'pending' and payment status 'requires_payment'
        - Create a payment intent with the selected payment provider (Stripe, PayPal, etc.)
        - Return payment details (client_secret, payment_intent_id) for client-side SDK integration
        - Keep cart items until payment is confirmed (cart cleared on payment confirmation)
        - Store complete address copies (not references) to preserve delivery information

        **Address Handling**: Complete address objects must be provided. If using saved addresses,
        the client should retrieve the full address details and include them in the request.
        This ensures immutable delivery/billing records that won't change if saved addresses are modified.

        The client should use the returned payment intent details with the appropriate
        payment SDK (Stripe Elements, PayPal SDK, etc.) to complete payment.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 20 requests per minute per user.
      operationId: createOrder
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              create_order:
                $ref: '#/components/examples/order_create_request'
      responses:
        '201':
          description: Order created successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
              examples:
                stripe_order_created:
                  $ref: '#/components/examples/stripe_order_created'
                paypal_order_created:
                  $ref: '#/components/examples/paypal_order_created'
        '400':
          description: Bad request - empty cart or invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                empty_cart:
                  $ref: '#/components/examples/empty_cart_2'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{id}:
    get:
      tags:
        - product
      summary: Get order details
      description: |
        Retrieve detailed information about a specific order.
        Only the order owner can access their order details.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 100 requests per minute per user.
      operationId: getOrder
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Order identifier
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Order details retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Access denied - not the order owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                access_denied:
                  $ref: '#/components/examples/access_denied_2'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/cancel:
    put:
      tags:
        - product
      summary: Cancel order
      description: |
        Cancel an order if it's still in a cancellable state (pending, confirmed).
        Orders that are already processing, shipped, or delivered cannot be cancelled.

        Requires valid JWT Bearer token in Authorization header.
        Rate limited to 10 requests per minute per user.
      operationId: cancelOrder
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Order identifier
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
            examples:
              cancel_order:
                $ref: '#/components/examples/cancel_order'
      responses:
        '200':
          description: Order cancelled successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderActionResponse'
              examples:
                order_cancelled:
                  $ref: '#/components/examples/order_cancelled'
        '400':
          description: Order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cannot_cancel:
                  $ref: '#/components/examples/cannot_cancel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Access denied - not the order owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                access_denied:
                  $ref: '#/components/examples/access_denied_3'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication

    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT Bearer token with elevated permissions

    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: |
        HTTP-only SameSite=Strict refresh token cookie. Set automatically on login, register,
        OAuth callback, and token refresh. Used by web clients to refresh access tokens without
        exposing the refresh token to JavaScript.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: |
        Opaque cursor token for pagination. Omit on the first request.
        Use the value from the `X-Next-Cursor` response header to fetch the next page.
        Currently the cursor is the ID of the last item in the previous page.
      schema:
        type: string
        example: "65f1a2b3c4d5e6f7g8h9i0j1"
  headers:
    RequestIdHeader:
      description: Unique request identifier for tracing and debugging
      schema:
        type: string
        pattern: '^req-[a-zA-Z0-9]+$'
        example: "req-abc123def456ghi789"
    NextCursorHeader:
      description: |
        Cursor token for the next page. Pass this as the `cursor` query parameter on the
        next request. Absent when there are no more items (`X-Has-More: false`).
        Currently the value is the ID of the last item in the returned list.
      schema:
        type: string
        example: "65f1a2b3c4d5e6f7g8h9i0j9"
...
  schemas:
...
    ID:
      type: string
      description: Identifier — can be numeric (Oracle sequence) or UUID v4
      example: "12345"
...